Add Neo4j in docker-compose
MongoDB won't be used to store relationships between costs (cost will be
renamed to composition). It's not efficient in querying nested data
(hierarchical trees).

Neo4j does a great job!

+++ unit/unit.go
@@ -3,10 +3,10 @@ package unit
 type Unit struct {
 	Type     string
 	Name     string
-	Modifier float32
+	Modifier float64
 }
 
-func NewUnit(t string, n string, m float32) *Unit {
+func NewUnit(t string, n string, m float64) *Unit {
 	return &Unit{
 		Type:     t,
 		Name:     n,
